[스터디 목록] 스터디 상세 화면 컴포넌트 분리 및 구조 개선
- 스터디 상세 화면 (`StudyDetailScreen.kt`)의 컴포넌트를 역할별로 분리했습니다.
- 스터디 헤더 (`StudyHeader.kt`), 스터디 정보 카드 (`StudyInfoCard.kt`), 카테고리 칩 (`CategoryChips.kt`), 스터디 설명 (`StudyDescription.kt`), 요구사항 (`StudyRequirements.kt`), 참가 버튼 (`JoinStudyButton`) 컴포넌트를 별도의 파일로 분리하여 가독성과 재사용성을 높였습니다.
- 분리된 컴포넌트들을 `StudyContent.kt`에 통합하여 상세 정보 내용을 구성했습니다.
- 네비게이션 경로를 새로운 패키지 구조에 맞게 업데이트했습니다.
