+++ app/src/main/java/com/example/smartee/ui/study/studyList/studydetail/StudyDescription.kt
@@ -1,0 +1,25 @@
+package com.example.smartee.ui.study.studyList.studydetail
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun StudyDescription(description: String) {
+    Text(
+        text = "스터디 소개",
+        style = MaterialTheme.typography.titleLarge,
+        fontWeight = FontWeight.Bold,
+        modifier = Modifier.padding(bottom = 8.dp)
+    )
+
+    Text(
+        text = description.ifEmpty { "스터디 설명이 없습니다." },
+        style = MaterialTheme.typography.bodyLarge,
+        modifier = Modifier.padding(bottom = 16.dp)
+    )
+}
