+++ app/src/main/java/com/example/smartee/ui/study/studyList/studydetail/StudyHeader.kt
@@ -1,0 +1,126 @@
+package com.example.smartee.ui.study.studyList.studydetail
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Flag
+import androidx.compose.material.icons.filled.MoreVert
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import coil3.compose.AsyncImage
+import com.example.smartee.model.StudyData
+
+@Composable
+fun StudyHeader(
+    study: StudyData,
+    onReportStudy: (String) -> Unit
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(250.dp)
+    ) {
+        // 배경 이미지
+        AsyncImage(
+            model = study.thumbnailModel,
+            contentDescription = "Study Thumbnail",
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.Crop
+        )
+
+        // 그라데이션 오버레이
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(
+                    Brush.verticalGradient(
+                        colors = listOf(
+                            Color.Transparent,
+                            Color.Black.copy(alpha = 0.7f)
+                        ),
+                        startY = 0f,
+                        endY = 500f
+                    )
+                )
+        )
+
+        // 옵션 메뉴 (더보기)
+        StudyOptionsMenu(study.studyId, onReportStudy)
+
+        // 제목 텍스트
+        Text(
+            text = study.title,
+            color = Color.White,
+            style = MaterialTheme.typography.headlineMedium,
+            fontWeight = FontWeight.Bold,
+            modifier = Modifier
+                .align(Alignment.BottomStart)
+                .padding(16.dp)
+        )
+    }
+}
+
+@Composable
+private fun StudyOptionsMenu(studyId: String, onReportStudy: (String) -> Unit) {
+    var showMenu by remember { mutableStateOf(false) }
+
+    IconButton(
+        onClick = { showMenu = true },
+        modifier = Modifier
+            .padding(8.dp)
+            .wrapContentSize(align = Alignment.TopEnd)
+    ) {
+        Icon(
+            imageVector = Icons.Default.MoreVert,
+            contentDescription = "More Options",
+            tint = Color.White
+        )
+    }
+
+    // 드롭다운 메뉴
+    DropdownMenu(
+        expanded = showMenu,
+        onDismissRequest = { showMenu = false }
+    ) {
+        androidx.compose.material3.DropdownMenuItem(
+            text = {
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Icon(
+                        imageVector = Icons.Default.Flag,
+                        contentDescription = "Report",
+                        tint = MaterialTheme.colorScheme.error
+                    )
+                    Spacer(modifier = Modifier.width(8.dp))
+                    Text("신고하기")
+                }
+            },
+            onClick = {
+                onReportStudy(studyId)
+                showMenu = false
+            }
+        )
+    }
+}
