+++ app/src/main/java/com/example/smartee/ui/study/studyList/studydetail/StudyInfoCard.kt
@@ -1,0 +1,128 @@
+package com.example.smartee.ui.study.studyList.studydetail
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.Comment
+import androidx.compose.material.icons.filled.CalendarToday
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material.icons.filled.LocationOn
+import androidx.compose.material.icons.filled.People
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+import com.example.smartee.model.StudyData
+
+@Composable
+fun StudyInfoCard(study: StudyData) {
+    Card(
+        modifier = Modifier.fillMaxWidth(),
+        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp),
+        shape = RoundedCornerShape(12.dp)
+    ) {
+        Column(
+            modifier = Modifier.padding(16.dp)
+        ) {
+            // 위치 정보
+            if (study.address.isNotEmpty()) {
+                InfoRow(
+                    icon = Icons.Default.LocationOn,
+                    text = study.address
+                )
+            }
+
+            // 인원 정보
+            InfoRow(
+                icon = Icons.Default.People,
+                text = if (study.maxMemberCount == Int.MAX_VALUE)
+                    "인원 제한 없음 (현재 ${study.currentMemberCount}명)"
+                else
+                    "${study.currentMemberCount}/${study.maxMemberCount}명"
+            )
+
+            // 날짜 정보
+            InfoRow(
+                icon = Icons.Default.CalendarToday,
+                text = study.getLocalDateTime().toString().split("T")[0] // yyyy-MM-dd 형식
+            )
+
+            // 통계 정보 (좋아요, 댓글 수)
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(16.dp)
+            ) {
+                StatCount(
+                    icon = Icons.Default.Favorite,
+                    count = study.likeCount,
+                    tint = MaterialTheme.colorScheme.error
+                )
+
+                StatCount(
+                    icon = Icons.AutoMirrored.Default.Comment,
+                    count = study.commentCount,
+                    tint = MaterialTheme.colorScheme.primary
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun InfoRow(
+    icon: ImageVector,
+    text: String,
+    tint: Color = MaterialTheme.colorScheme.primary
+) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = Modifier.padding(bottom = 8.dp)
+    ) {
+        Icon(
+            imageVector = icon,
+            contentDescription = null,
+            tint = tint,
+            modifier = Modifier.size(20.dp)
+        )
+        Spacer(modifier = Modifier.width(8.dp))
+        Text(
+            text = text,
+            style = MaterialTheme.typography.bodyLarge
+        )
+    }
+}
+
+@Composable
+private fun StatCount(
+    icon: ImageVector,
+    count: Int,
+    tint: Color
+) {
+    Row(verticalAlignment = Alignment.CenterVertically) {
+        Icon(
+            imageVector = icon,
+            contentDescription = null,
+            tint = tint,
+            modifier = Modifier.size(20.dp)
+        )
+        Spacer(modifier = Modifier.width(4.dp))
+        Text(
+            text = "$count",
+            style = MaterialTheme.typography.bodyLarge
+        )
+    }
+}
